Move set_key_release_policy into application code
