+++ app.json
@@ -48,7 +48,7 @@
         "js_function": "keyReleasePolicy",
         "forwarding_required": "sometimes",
         "redirection_strategy": "none",
-        "authn_policies": ["member_cert"],
+        "authn_policies": [],
         "mode": "readonly",
         "openapi": {
           "responses": {
